feat: add styles
