--- virtual-keyboard/css/fonts.css
@@ -1,1 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Open+Sans&display=swap');
