+++ virtual-keyboard/css/style.css
@@ -1,113 +1,127 @@
-
 * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: 'Open Sans', sans-serif;
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: 'Open Sans', sans-serif;
+  
 }
 
-h1 {
-    display: block;
-    font-size: 16px;
-    margin-block-start: 0.67em;
-    margin-block-end: 0.67em;
-    margin-inline-start: 0px;
-    margin-inline-end: 0px;
-    font-weight: bold;
-}
-
-.wrapper {
-    max-width: 816px;
-    margin: 0 auto;
-}
-
-textarea {
-    width: 100%;
-    height: 30vh;
-    font-size: 20px;
-    resize: vertical;
-    border-radius: 5px;
+main {
+  width: 1020px;
+  margin: 0 auto;
 }
 
 .keyboard {
-    width: 100%;
-    margin-top: 30px;
-    max-width: 816px;
-    padding: 10px 8px 8px;
-    margin: 30px auto;
-    background-color: #e3e3e3;
-    border-radius: 5px;
+  display: grid;
+  grid-row-gap: 10px;
+  margin: 20px auto;
+  padding: 10px 8px 8px;
+  border: 1px solid;
+  position: relative;
+  overflow: hidden;
+  background-color: #e3e3e3;
+  border-radius: 5px;
+  border: none;
 }
 
-.line {
-    display: flex;
-    justify-content: space-between;
+.keyboard__row {
+  display: grid;
+  grid-gap: 10px;
+  grid-row-gap: 10px;
+  height: 50px;
 }
 
 .keyboard__key {
-    position: relative;
-    width: 5%;
-    max-width: 50px;
-    height: 50px;
-    margin: 3px;
-    font-size: 16px;
-    color: black;
-    cursor: pointer;
-    border: none;
-    border-radius: 5px;
-    outline: none;
-    box-shadow: inset 2px 2px 2px 0px rgb(255 255 255 / 50%), 7px 7px 20px 0px rgb(0 0 0 / 10%), 4px 4px 5px 0px rgb(0 0 0 / 10%);
-    transition: box-shadow 0.3s, background-color 0.3s;
+  display: flex;
+  cursor: pointer;
+  border-radius: 5px;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  box-shadow: inset 2px 2px 2px 0px rgb(255 255 255 / 50%), 7px 7px 20px 0px rgb(0 0 0 / 10%), 4px 4px 5px 0px rgb(0 0 0 / 10%);
+  transition: box-shadow 0.3s, background-color 0.3s;
 }
 
-.keyboard__key:active {
-    background-color: rgba(255,151, 0, 0.2);
-    box-shadow: 0 0 1px rgba(0, 0, 0, 0.5);
+[data-code="CapsLock"] {
+  width: 120px;
 }
 
-.keyboard__key_wide {
-    width: 8%;
-    max-width: 75px;
+[data-code="ShiftLeft"] {
+  width: 100px;
 }
 
-.keyboard__key_middle-wide {
-    width: 15%;
-    max-width: 150px;
+[data-code="ShiftRight"] {
+  width: 120px;
 }
 
-.keyboard__key_ultra-wide {
-    width: 33%;
-    max-width: 334px;
+[data-code="Enter"] {
+  width: 100px;
 }
 
-@media (hover: hover) {
-    .keyboard__key:hover {
-        background: linear-gradient(0deg, rgba(255,151,0,1) 0%, rgba(251,75,2,1) 100%)
-    }
+[data-code="Tab"] {
+  width: 76px;
+}
+
+[data-code="ControlLeft"] {
+  width: 76px;
+}
+
+[data-code="ControlRight"] {
+  width: 60px;
+}
+
+[data-code="Backspace"] {
+  width: 115px;
+}
+
+[data-code="Space"] {
+  width: 420px;
+}
+
+[data-code="Delete"] {
+  width: 100px;
+}
+
+[data-code="AltLeft"],
+[data-code="AltRight"] {
+  width: 76px;
 }
 
 .active {
-    background: linear-gradient(0deg, rgba(255,151,0,1) 0%, rgba(251,75,2,1) 100%);
-    box-shadow: 0 0 1px rgba(0, 0, 0, 0.5);
+  background: linear-gradient(0deg, rgba(255,151,0,1) 0%, rgba(251,75,2,1) 100%);
+  box-shadow: 0 0 1px rgba(0, 0, 0, 0.5);
 }
 
-.shift {
-    text-transform: uppercase;
+.letter {
+  font-size: 20px;
+  width: 100%;
+  text-align: center;
+  transition: all .2s ease;
 }
 
-.always-shift {
-    position: absolute;
-    top: 0;
-    right: 0;
-    width: 30px;
-    height: 30px;
-    padding: 0;
-    margin: 0;
-    border: none;
-    border-radius: 5px;
+h1, h3 {
+  color: #686886;
+  margin-top: 20px;
+  margin-bottom: 20px;
+  text-align: center;
+  font-weight: 400;
 }
 
-.always-shift::after {
-    font-size: 18px;
-    content: '^';
+textarea.output {
+  resize: vertical;
+  border-radius: 5px;
+  width: 100%;
+  height: 30vh;
+  font-size: 20px;
+  padding: 20px 8px 8px;
+  box-sizing: border-box;
+  letter-spacing: 1px;
+  font-weight: 400;
+}
+
+.sub {
+  width: 100%;
+  color: #686886;
+  text-align: center;
+  width: 50%;
 }
