+++ virtual-keyboard/index.html
@@ -5,9 +5,9 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Virtual Keyboard</title>
-    <link rel="stylesheet" href=".css/style.css">
+    <link rel="stylesheet" href="css/style.css">
 </head>
 <body>
-    <script src=".js/script.js" type="module"></script>
+    <script src="js/script.js" type="module"></script>
 </body>
 </html>
