+++ virtual-keyboard/js/Key.js
@@ -1,6 +1,6 @@
 /* eslint-disable import/extensions */
 
-import create from '.js/utils/create.js';
+import create from './utils/create.js';
 
 export default class Key {
 constructor( {small, code, shift} ) {
