+++ virtual-keyboard/js/layouts/index.js
@@ -1,6 +1,6 @@
 /* eslint-disable import/extensions */
-import ru from '.js/layouts/ru.js';
-import en from '.js/layouts/en.js';
+import ru from './ru.js';
+import en from './en.js';
 
 export default { ru, en };
