+++ virtual-keyboard/js/script.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/extensions */
 
-import { get } from '.js/utils/storage.js';
-import Keyboard from '.js/utils/Keyboard.js';
+import { get } from './storage.js';
+import Keyboard from './Keyboard.js';
 
 const rowsOrder = [
     ['Backquote', 'Digit1', 'Digit2', 'Digit3', 'Digit4', 'Digit5', 'Digit6', 'Digit7', 'Digit8', 'Digit9', 'Digit0', 'Minus', 'Equal', 'Delete'],
