+++ virtual-keyboard/js/storage.js
